backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
python-docx==1.1.2
PyPDF2==3.0.1
httpx==0.27.2
resend==2.4.0
python-dotenv==1.0.1
